pyproject.toml
[project]
name = "rpa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anthropic==0.84.0",
    "anyio==4.12.1",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "comtypes==1.4.15",
    "cryptography==46.0.5",
    "distro==1.9.0",
    "docstring-parser==0.17.0",
    "filelock==3.24.3",
    "fsspec==2026.2.0",
    "google-auth==2.48.0",
    "google-genai==1.65.0",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "hf-xet==1.3.1",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==1.4.1",
    "idna==3.11",
    "jiter==0.13.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.2.10",
    "langchain-core==1.2.16",
    "langchain-openai==1.1.10",
    "langgraph==1.0.9",
    "langgraph-checkpoint==4.0.0",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.9",
    "langsmith==0.7.7",
    "markdown-it-py==4.0.0",
    "mdurl==0.1.2",
    "mouseinfo==0.1.3",
    "numpy==2.4.2",
    "openai==2.24.0",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==26.0",
    "playwright==1.58.0",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.2",
    "pyautogui==0.9.54",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pyee==13.0.1",
    "pygetwindow==0.0.9",
    "pygments==2.19.2",
    "pymsgbox==2.0.1",
    "pyperclip==1.11.0",
    "pyrect==0.2.0",
    "pyscreeze==1.0.1",
    "python-dotenv==1.2.1",
    "pytweening==1.2.0",
    "pywin32==311",
    "pywinauto==0.6.9",
    "pyyaml==6.0.3",
    "regex==2026.2.19",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.3.3",
    "rsa==4.9.1",
    "safetensors==0.7.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "tenacity==9.1.4",
    "tiktoken==0.12.0",
    "tokenizers==0.22.2",
    "tqdm==4.67.3",
    "transformers==5.2.0",
    "typer==0.24.1",
    "typer-slim==0.24.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "websockets==16.0",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
]
